requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# LLM & Agents
anthropic==0.34.2
langchain==0.3.1
langchain-text-splitters==0.3.0
langfuse==2.53.0

# Embeddings
sentence-transformers==3.1.1

# Vector DB
qdrant-client==1.11.3

# Database
aiosqlite==0.20.0
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3

# HTTP
httpx==0.27.2

# Utilities
structlog==24.4.0